Added FusionMenuListener module
